// src/app/common/data-core/enums/geo/map-element-type.enum.ts
/**	MapElementType (元素类型) v1.1	*/
export enum MapElementType {
  /**	区域	*/
  Range = 1,
  /**	楼栋	*/
  Building = 2,
  /**	房屋	*/
  House = 3,
  /**	摄像机	*/
  Camera = 4,
  /**	门	*/
  Door = 5,
  /**	出入口	*/
  Entrance = 6,
  /**	楼层	*/
  Floor = 7,
  /**	传感器	*/
  IoTSensor = 8,
  /**	报警器	*/
  Announciator = 9,
}